creating new request for get user's list, in this component don't need more logic
